regroup uxd itxs form into UXD folder
